QPID-6021: fix compile error on windows
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1619318 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qpid/broker/amqp/Connection.cpp
@@ -168,9 +168,9 @@ size_t Connection::encode(char* buffer, size_t size)
     }
 }
 
-void Connection::doOutput(ssize_t capacity)
+void Connection::doOutput(size_t capacity)
 {
-    for (ssize_t n = pn_transport_pending(transport); n > 0 && n < capacity; n = pn_transport_pending(transport)) {
+    for (ssize_t n = pn_transport_pending(transport); n > 0 && n < (ssize_t) capacity; n = pn_transport_pending(transport)) {
         if (dispatch()) processDeliveries();
         else break;
     }
